Create new automation image 10-30-25
Signed-off-by: Tim Zhou <tzhou@redhat.com>

Specify WiX Toolset version on Windows

The WiX version should be consistent with the one that is specified to
build the Podman installer.

Moreover the WiX toolset licensing has changed with the release of
version 6 and we should not upgrade to it unless Podman organization
becomes a sponsor of the project:

https://github.com/containers/podman/issues/27042

Signed-off-by: Mario Loriedo <mario.loriedo@gmail.com>

skopeo_cidev: add docker-distribution

Signed-off-by: Lokesh Mandvekar <lsm5@redhat.com>

Install packages for protoc-gen-go and protoc-gen-go-grpc

Install the packages that provide the protoc protobuf compiler with the
ability to generate Go and Go GRPC support.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>